+++ auth-filters/forgerock-authn-filter/forgerock-jaspi-functional-tests/src/main/java/org/forgerock/caf/authn/test/runtime/GuiceModule.java
@@ -20,10 +20,9 @@ import com.google.inject.AbstractModule;
 import com.google.inject.Injector;
 import com.google.inject.Provides;
 import org.forgerock.caf.authn.test.configuration.ConfigurationResource;
+import org.forgerock.jaspi.runtime.ContextFactory;
 import org.forgerock.jaspi.context.FallbackServerAuthContext;
 import org.forgerock.jaspi.runtime.AuditApi;
-import org.forgerock.jaspi.runtime.JaspiRuntime;
-import org.forgerock.jaspi.runtime.RuntimeResultHandler;
 import org.forgerock.jaspi.runtime.context.ContextHandler;
 import org.forgerock.jaspi.utils.MessageInfoUtils;
 import org.forgerock.json.fluent.JsonValue;
@@ -58,8 +57,7 @@ public class GuiceModule extends AbstractModule {
     @Override
     protected void configure() {
         bind(MessageInfoUtils.class).in(Singleton.class);
-        bind(RuntimeResultHandler.class).in(Singleton.class);
-        bind(ServerAuthContext.class).to(ConfigurableServerAuthContext.class).in(Singleton.class);
+        bind(ContextFactory.class).to(TestContextFactory.class);
         bind(AuditApi.class).to(TestAuditApi.class);
     }
 
@@ -74,22 +72,6 @@ public class GuiceModule extends AbstractModule {
     @Singleton
     public ContextHandler getContextHandler(MessageInfoUtils messageInfoUtils) {
         return new ContextHandler(messageInfoUtils);
-    }
-
-    /**
-     * Provider for the JASPI runtime instance.
-     *
-     * @param serverAuthContext The {@code ServerAuthContext} instance.
-     * @param resultHandler An instance of the {@code RuntimeResultHandler}.
-     * @param auditApi An instance of the {@code AuditApi}.
-     * @return The JASPI runtime instance.
-     */
-    @Provides
-    @Inject
-    @Singleton
-    public JaspiRuntime getJaspiRuntime(ServerAuthContext serverAuthContext,
-            RuntimeResultHandler resultHandler, AuditApi auditApi) {
-        return new JaspiRuntime(serverAuthContext, resultHandler, auditApi);
     }
 
     /**